chore: rename dbutil package to store
