+++ api/pkg/cmd/cmd_test.go
@@ -5,9 +5,9 @@ import (
 	"sync"
 	"testing"
 
-	"github.com/chanzuckerberg/happy/api/pkg/dbutil"
 	"github.com/chanzuckerberg/happy/api/pkg/ent/enttest"
 	"github.com/chanzuckerberg/happy/api/pkg/setup"
+	"github.com/chanzuckerberg/happy/api/pkg/store"
 	"github.com/google/uuid"
 )
 
@@ -15,7 +15,7 @@ var (
 	mu sync.Mutex
 )
 
-func MakeTestDB(t *testing.T) *dbutil.DB {
+func MakeTestDB(t *testing.T) *store.DB {
 	config := setup.GetConfiguration()
 
 	// Even with a UUID in the data source name this is not thread safe so we need to use a mutex to prevent concurrent access
@@ -23,5 +23,5 @@ func MakeTestDB(t *testing.T) *dbutil.DB {
 	client := enttest.Open(t, "sqlite3", fmt.Sprintf("file:memdb%s?mode=memory&cache=shared&_fk=1", uuid.NewString()))
 	mu.Unlock()
 
-	return dbutil.MakeDB(config.Database).WithClient(client)
+	return store.MakeDB(config.Database).WithClient(client)
 }
